+++ test/command/test_scan_policy_file.py
@@ -8,78 +8,59 @@ from cloudsplaining.shared.exclusions import DEFAULT_EXCLUSIONS, Exclusions
 
 class PolicyFileTestCase(unittest.TestCase):
     def test_policy_file(self):
-        policy_test_file = os.path.abspath(
-            os.path.join(
-                os.path.dirname(__file__),
-                os.path.pardir,
-                "files",
-                "test_policy_file.json",
-            )
-        )
-
-        # print(expected_results_file)
-        with open(policy_test_file) as json_file:
-            example_policy = json.load(json_file)
+        example_policy = {
+            "Version": "2012-10-17",
+            "Statement": [
+                {
+                    "Effect": "Allow",
+                    "Action": [
+                        "ecr:GetAuthorizationToken",
+                        "ecr:BatchCheckLayerAvailability",
+                        "ecr:GetDownloadUrlForLayer",
+                        "ecr:GetRepositoryPolicy",
+                        "ecr:DescribeRepositories",
+                        "ecr:ListImages",
+                        "ecr:DescribeImages",
+                        "ecr:BatchGetImage",
+                        "ecr:InitiateLayerUpload",
+                        "ecr:UploadLayerPart",
+                        "ecr:CompleteLayerUpload",
+                        "ecr:PutImage"
+                    ],
+                    "Resource": "*"
+                },
+              {
+                    "Sid": "AllowManageOwnAccessKeys",
+                    "Effect": "Allow",
+                    "Action": [
+                        "iam:CreateAccessKey",
+                        "iam:DeleteAccessKey",
+                        "iam:ListAccessKeys",
+                        "iam:UpdateAccessKey"
+                    ],
+                    "Resource": "arn:aws:iam::*:user/${aws:username}"
+                }
+            ]
+        }
         expected_results = {
-            "AccountID": "N/A",
-            "ManagedBy": "Customer",
-            "PolicyName": "test",
-            "Name": "test",
-            "Type": "",
-            "Arn": "test",
-            "AttachedToPrincipal": None,
-            "ActionsCount": 4,
-            "ServicesCount": 1,
-            "Services": [
+            "ServicesAffected": [
                 "ecr"
             ],
-            "Actions": [
+            "PrivilegeEscalation": [],
+            "ResourceExposure": [],
+            "DataExfiltration": [],
+            "ServiceWildcard": [],
+            "CredentialsExposure": [
+                "ecr:GetAuthorizationToken"
+            ],
+            "InfrastructureModification": [
                 "ecr:CompleteLayerUpload",
                 "ecr:InitiateLayerUpload",
                 "ecr:PutImage",
                 "ecr:UploadLayerPart"
-            ],
-            "PolicyDocument": {
-                "Version": "2012-10-17",
-                "Statement": [
-                    {
-                        "Effect": "Allow",
-                        "Action": [
-                            "ecr:GetAuthorizationToken",
-                            "ecr:BatchCheckLayerAvailability",
-                            "ecr:GetDownloadUrlForLayer",
-                            "ecr:GetRepositoryPolicy",
-                            "ecr:DescribeRepositories",
-                            "ecr:ListImages",
-                            "ecr:DescribeImages",
-                            "ecr:BatchGetImage",
-                            "ecr:InitiateLayerUpload",
-                            "ecr:UploadLayerPart",
-                            "ecr:CompleteLayerUpload",
-                            "ecr:PutImage"
-                        ],
-                        "Resource": "*"
-                    },
-                    {
-                        "Sid": "AllowManageOwnAccessKeys",
-                        "Effect": "Allow",
-                        "Action": [
-                            "iam:CreateAccessKey",
-                            "iam:DeleteAccessKey",
-                            "iam:ListAccessKeys",
-                            "iam:UpdateAccessKey"
-                        ],
-                        "Resource": "arn:aws:iam::*:user/${aws:username}"
-                    }
-                ]
-            },
-            "AssumeRolePolicyDocument": None,
-            "AssumableByComputeService": [],
-            "PrivilegeEscalation": [],
-            "DataExfiltration": [],
-            "ResourceExposure": []
+            ]
         }
-        results = scan_policy(example_policy, "test", DEFAULT_EXCLUSIONS)
+        results = scan_policy(example_policy)
         # print(json.dumps(results, indent=4))
         self.maxDiff = None
         self.assertDictEqual(results, expected_results)
@@ -99,40 +80,12 @@ class PolicyFileTestCase(unittest.TestCase):
                 },
             ]
         }
-        results = scan_policy(test_policy, "test", DEFAULT_EXCLUSIONS)
-        expected_results_before_exclusion = {
-            "AccountID": "N/A",
-            "ManagedBy": "Customer",
-            "PolicyName": "test",
-            "Name": "test",
-            "Type": "",
-            "Arn": "test",
-            "AttachedToPrincipal": None,
-            "ActionsCount": 2,
-            "ServicesCount": 2,
-            "Services": [
+        results = scan_policy(test_policy)
+        expected_results = {
+            "ServicesAffected": [
                 "iam",
                 "s3"
             ],
-            "Actions": [
-                "iam:CreateAccessKey",
-                "s3:GetObject"
-            ],
-            "PolicyDocument": {
-                "Version": "2012-10-17",
-                "Statement": [
-                    {
-                        "Effect": "Allow",
-                        "Action": [
-                            "s3:GetObject",
-                            "iam:CreateAccessKey"
-                        ],
-                        "Resource": "*"
-                    }
-                ]
-            },
-            "AssumeRolePolicyDocument": None,
-            "AssumableByComputeService": [],
             "PrivilegeEscalation": [
                 {
                     "type": "CreateAccessKey",
@@ -141,16 +94,24 @@ class PolicyFileTestCase(unittest.TestCase):
                     ]
                 }
             ],
+            "ResourceExposure": [
+                "iam:CreateAccessKey"
+            ],
             "DataExfiltration": [
                 "s3:GetObject"
             ],
-            "ResourceExposure": [
+            "ServiceWildcard": [],
+            "CredentialsExposure": [
                 "iam:CreateAccessKey"
+            ],
+            "InfrastructureModification": [
+                "iam:CreateAccessKey",
+                "s3:GetObject"
             ]
         }
-        # print(json.dumps(results, indent=4))
+        print(json.dumps(results, indent=4))
         self.maxDiff = None
-        self.assertDictEqual(results, expected_results_before_exclusion)
+        self.assertDictEqual(results, expected_results)
 
     def test_excluded_actions_scan_policy_file_v2(self):
         """test_excluded_actions_scan_policy_file_v2: Test the scan_policy command when we have excluded actions"""
@@ -167,18 +128,35 @@ class PolicyFileTestCase(unittest.TestCase):
                 },
             ]
         }
-        expected_results_after_exclusion = {}
-        exclusions_cfg_custom = {
-            "users": ["MyRole"],
-            "groups": ["obama"],
-            "roles": ["admin"],
-            "exclude-actions": [
-                "s3:GetObject",
+        expected_results = {
+            "ServiceWildcard": [],
+            "ServicesAffected": [
+                "iam",
+                "s3"
+            ],
+            "PrivilegeEscalation": [
+                {
+                    "type": "CreateAccessKey",
+                    "actions": [
+                        "iam:createaccesskey"
+                    ]
+                }
+            ],
+            "ResourceExposure": [
+                "iam:CreateAccessKey"
+            ],
+            "DataExfiltration": [
+                "s3:GetObject"
+            ],
+            "CredentialsExposure": [
+                "iam:CreateAccessKey"
+            ],
+            "InfrastructureModification": [
                 "iam:CreateAccessKey"
             ]
         }
-        exclusions = Exclusions(exclusions_cfg_custom)
-        results = scan_policy(test_policy, "test", exclusions)
+        exclusions_cfg_custom = {}
+        results = scan_policy(test_policy, exclusions_cfg_custom)
         # print(json.dumps(results, indent=4))
         self.maxDiff = None
-        self.assertDictEqual(results, expected_results_after_exclusion)
+        self.assertDictEqual(results, expected_results)